build: Reorganize directory structure
* Move header files, library implementation, and example agents all to
  separate directories
* Move CMake code to separate CMakeLists.txt files in the new subfolders
* Rename mosqagent.[ch] files to mqtta.[ch]
