+++ CMakeLists.txt
@@ -33,76 +33,13 @@ find_library(CONFIG_LIBRARY NAMES config)
 find_library(MOSQUITTO_LIBRARY NAMES mosquitto)
 
 # Headers
-configure_file(
-    mqtta-build.h.in
-    mqtta-build.h
-    @ONLY
-)
-include_directories(
-    "${CMAKE_CURRENT_BINARY_DIR}"
-)
+add_subdirectory(include/mqtt-tools)
 
 # Libraries
-
-## mosqhelper
-add_library(mosqhelper
-    mosqhelper.c
-)
-set_target_properties(mosqhelper PROPERTIES
-    OUTPUT_NAME mosqhelper
-    SOVERSION   ${PROJECT_VERSION_MAJOR}
-    VERSION     ${PROJECT_VERSION}
-)
-target_link_libraries(mosqhelper
-    "${MOSQUITTO_LIBRARY}"
-)
-install(TARGETS mosqhelper
-	EXPORT ${PROJECT_NAME}-targets
-	LIBRARY DESTINATION "${CMAKE_INSTALL_LIBDIR}"
-	ARCHIVE DESTINATION "${CMAKE_INSTALL_LIBDIR}"
-)
-install(FILES
-	mosqhelper.h
-	DESTINATION "${CMAKE_INSTALL_INCLUDEDIR}/${PROJECT_NAME}"
-)
-
-## mosqagent
-add_library(mosqagent
-    mosqagent.c
-)
-set_target_properties(mosqagent PROPERTIES
-    OUTPUT_NAME mosqagent
-    SOVERSION   ${PROJECT_VERSION_MAJOR}
-    VERSION     ${PROJECT_VERSION}
-)
-target_link_libraries(mosqagent
-    mosqhelper
-    "${CONFIG_LIBRARY}"
-    "${MOSQUITTO_LIBRARY}"
-)
-install(TARGETS mosqagent
-	EXPORT ${PROJECT_NAME}-targets
-	LIBRARY DESTINATION "${CMAKE_INSTALL_LIBDIR}"
-	ARCHIVE DESTINATION "${CMAKE_INSTALL_LIBDIR}"
-)
-install(FILES
-	mosqagent.h
-	DESTINATION "${CMAKE_INSTALL_INCLUDEDIR}/${PROJECT_NAME}"
-)
+add_subdirectory(lib)
 
 # Agents
-
-# mqtt-clock
-add_executable(mqtt-clock
-    mqtt-clock.c
-)
-target_link_libraries(mqtt-clock
-    mosqhelper
-    mosqagent
-)
-install(TARGETS mqtt-clock
-	RUNTIME DESTINATION "${CMAKE_INSTALL_BINDIR}"
-)
+add_subdirectory(agents)
 
 # Install CMake targets
 install(EXPORT ${PROJECT_NAME}-targets
