+++ include/mqtt-tools/CMakeLists.txt
@@ -1,0 +1,11 @@
+configure_file(
+    mqtta-build.h.in
+    mqtta-build.h
+    @ONLY
+)
+
+install(FILES
+	mqtta.h
+	mosqhelper.h
+	DESTINATION "${CMAKE_INSTALL_INCLUDEDIR}/mqtt-tools"
+)
