+++ lib/mosqhelper.c
